Extract eframe::App impl to frame_update.rs (refs #68)
Moves the entire `impl eframe::App for AmuxApp` block (update() +
on_exit()) out of main.rs into a dedicated frame_update.rs module.
update() is the ~350-line per-frame orchestrator that drains IPC,
handles input, lays out the sidebar + tabs + panes, renders modals,
and schedules repaint. on_exit() saves or clears the session.

Mirrors wezterm-gui's termwindow/render/paint.rs — the top-level
per-frame entry that dispatches to per-concern helpers.

main.rs drops from 843 to 475 LoC.

No behavior change — pure mechanical refactor.
